Issue #161 - Fix Controller name
